+++ src/blocks/Hero/__stories__/Hero.stories.tsx
@@ -151,10 +151,12 @@ VerticalOffset.parameters = {
     },
 };
 
-Theme.args = [data.themeVariants.light, data.themeVariants.dark].map((variant) => ({
-    ...data.default,
-    ...variant,
-})) as VariantsTemplateModel;
+Theme.args = [data.themeVariants.default, data.themeVariants.light, data.themeVariants.dark].map(
+    (variant) => ({
+        ...data.default,
+        ...variant,
+    }),
+) as VariantsTemplateModel;
 Theme.parameters = {
     controls: {
         include: Object.keys(Theme.args),