+++ src/components/Title/Title.tsx
@@ -1,6 +1,6 @@
 import {Col, GridColumnSizesType, GridJustifyContent} from '../../gravity-blocks/grid';
-import {ClassNameProps, TitleItemProps, TitleProps as TitleParams} from '../../models';
-import {block} from '../../utils';
+import {ClassNameProps, QAProps, TitleItemProps, TitleProps as TitleParams} from '../../models';
+import {block, getQaAttrubutes} from '../../utils';
 import YFMWrapper from '../YFMWrapper/YFMWrapper';
 
 import TitleItem from './TitleItem';
@@ -9,7 +9,7 @@ import './Title.scss';
 
 const b = block('title');
 
-export interface TitleProps extends TitleParams {
+export interface TitleProps extends TitleParams, QAProps {
     colSizes?: GridColumnSizesType;
     colJustifyContent?: GridJustifyContent;
     id?: string;
@@ -22,10 +22,13 @@ const Title = ({
     colSizes = {all: 12, sm: 8},
     colJustifyContent,
     id,
+    qa,
 }: TitleProps & ClassNameProps) => {
     if (!title && !subtitle) {
         return null;
     }
+
+    const qaAttributes = getQaAttrubutes(qa, ['title', 'subtitle']);
 
     const {text, ...titleProps} =
         !title || typeof title === 'string' ? ({text: title} as TitleItemProps) : title;
@@ -38,7 +41,7 @@ const Title = ({
                     sizes={colSizes}
                     {...(colJustifyContent && {justifyContent: colJustifyContent})}
                 >
-                    <TitleItem text={text} {...titleProps} />
+                    <TitleItem text={text} {...titleProps} qa={qaAttributes.title} />
                 </Col>
             )}
             {subtitle && (
@@ -47,7 +50,10 @@ const Title = ({
                     sizes={colSizes}
                     {...(colJustifyContent && {justifyContent: colJustifyContent})}
                 >
-                    <div className={b('description', {titleSize: titleProps?.textSize})}>
+                    <div
+                        className={b('description', {titleSize: titleProps?.textSize})}
+                        data-qa={qaAttributes.subtitle}
+                    >
                         <YFMWrapper content={subtitle} modifiers={{constructor: true}} />
                     </div>
                 </Col>