+++ src/gravity-blocks/grid/types.ts
@@ -53,7 +53,7 @@ export interface GridColumnClassParams {
     offsets?: GridColumnSizesType;
     orders?: GridColumnOrderSizesType;
     hidden?: GridColumnSize;
-    visible?: GridColumnSize;
+    visible?: GridColumnSize | Partial<Record<GridColumnSize, boolean>>;
     alignSelf?: GridColumnAlignSelf;
     justifyContent?: GridJustifyContent;
     reset?: boolean;